requirements.txt
pillow
aiohttp
python-dotenv
pytest
pytest-asyncio